remove all ui changes I guess
