+++ ui/src/artist/ArtistShow.jsx
@@ -1,11 +1,5 @@
 import React, { useState, createElement, useEffect } from 'react'
-import {
-  makeStyles,
-  MenuItem,
-  Select,
-  useMediaQuery,
-  withWidth,
-} from '@material-ui/core'
+import { useMediaQuery, withWidth } from '@material-ui/core'
 import {
   useShowController,
   ShowContextProvider,
@@ -13,21 +7,12 @@ import {
   useShowContext,
   ReferenceManyField,
   Pagination,
-  useTranslate,
 } from 'react-admin'
 import subsonic from '../subsonic'
 import AlbumGridView from '../album/AlbumGridView'
 import MobileArtistDetails from './MobileArtistDetails'
 import DesktopArtistDetails from './DesktopArtistDetails'
 import { useAlbumsPerPage } from '../common/index.js'
-import { useArtistRoles } from '../common/useArtistRoles.jsx'
-import { useLocation } from 'react-router-dom/cjs/react-router-dom.min.js'
-
-const useStyles = makeStyles({
-  root: {
-    padding: '1em',
-  },
-})
 
 const ArtistDetails = (props) => {
   const record = useRecordContext(props)
@@ -66,67 +51,39 @@ const ArtistDetails = (props) => {
 }
 
 const AlbumShowLayout = (props) => {
-  const translate = useTranslate()
   const showContext = useShowContext(props)
   const record = useRecordContext()
   const { width } = props
   const [, perPageOptions] = useAlbumsPerPage(width)
-  const { search } = useLocation()
-  const [role, setRole] = useState(
-    new URLSearchParams(search).get('role') ?? 'total',
-  )
-  const roles = useArtistRoles(false)
-  const classes = useStyles()
 
-  const maxPerPage = 36
+  const maxPerPage = 90
   let perPage = 0
   let pagination = null
 
-  const count =
-    role === 'total'
-      ? record?.albumCount
-      : record?.stats?.[role]?.albumCount || 0
+  const count = Math.max(
+    record?.stats?.['albumartist']?.albumCount || 0,
+    record?.stats?.['artist']?.albumCount ?? 0,
+  )
 
   if (count > maxPerPage) {
     perPage = Math.trunc(maxPerPage / perPageOptions[0]) * perPageOptions[0]
     const rowsPerPageOptions = [1, 2, 3].map((option) =>
-      Math.trunc(option * perPage),
+      Math.trunc(option * (perPage / 3)),
     )
     pagination = <Pagination rowsPerPageOptions={rowsPerPageOptions} />
   }
 
-  const id = `role_${role}_id`
-
   return (
     <>
       {record && <ArtistDetails />}
-      <div className={classes.root}>
-        <Select
-          value={role}
-          onChange={(event) => setRole(event.target.value)}
-          fullWidth
-        >
-          <MenuItem key="total" value="total">
-            {translate('resources.artist.fields.allRoles')} (
-            {record?.albumCount || 0})
-          </MenuItem>
-          {roles
-            .filter((role) => record?.stats?.[role.id]?.albumCount)
-            .map((role) => (
-              <MenuItem key={role.id} value={role.id}>
-                {role.name} ({record.stats[role.id].albumCount})
-              </MenuItem>
-            ))}
-        </Select>
-      </div>
       {record && (
         <ReferenceManyField
           {...showContext}
           addLabel={false}
           reference="album"
-          target={id}
+          target="artist_id"
           sort={{ field: 'max_year', order: 'ASC' }}
-          filter={{ [id]: record?.id }}
+          filter={{ artist_id: record?.id }}
           perPage={perPage}
           pagination={pagination}
         >
